+++ .gitignore
@@ -24,3 +24,4 @@ web-candidate-tracker/.gitignore
 
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
+/web-candidate-tracker/target/
