+++ web-candidate-tracker/pom.xml
@@ -35,9 +35,12 @@
             <artifactId>mysql-connector-java</artifactId>
             <version>8.0.28</version>
         </dependency>
-         
-
-
+	 <dependency>
+	  <groupId>jakarta.annotation</groupId>
+	  <artifactId>jakarta.annotation-api</artifactId>
+	  <version>1.3.5</version>
+	  <type>jar</type>
+	 </dependency>
 	</dependencies>
 
 	<build>
@@ -70,5 +73,6 @@
 
 		</plugins>
 	</build>
+    <name>maven-web-candidates-tracker</name>
 </project>
