+++ web-candidate-tracker/src/main/java/com/code/web/TestServlet.java
@@ -1,4 +1,4 @@
-package com.luv2code.web.jdbc;
+package com.code.web;
 
 import java.io.IOException;
 import java.io.PrintWriter;
