+++ web-candidate-tracker/src/main/webapp/META-INF/context.xml
@@ -1,9 +1,4 @@
-<Context>
-
-  <Resource name="jdbc/web_candidates" 
-  			auth="Container" type="javax.sql.DataSource"
-               maxActive="20" maxIdle="5" maxWait="10000"
-               username="recruiter" password="recruiter" 
-               driverClassName="com.mysql.cj.jdbc.Driver"
-               url="jdbc:mysql://localhost:3306/web_candidates?useSSL=false&amp;serverTimezone=UTC"/>
+<?xml version="1.0" encoding="UTF-8"?>
+<Context path="/web_candidates">
+  <Resource auth="Container" driverClassName="com.mysql.cj.jdbc.Driver" maxActive="20" maxIdle="5" maxWait="10000" name="jdbc/web_candidates" password="recruiter" type="javax.sql.DataSource" url="jdbc:mysql://localhost:3306/web_candidates?useSSL=false&amp;serverTimezone=UTC" username="recruiter"/>
 </Context>
