+++ web-candidate-tracker/src/main/webapp/update-candidate-form.jsp
@@ -18,25 +18,25 @@
 			<h3 class="title-form">Update Candidate</h3>
 			
 			<input type="hidden" name="command" value="UPDATE" />
-			<input type="hidden" name="studentId" value="${THE_STUDENT.id}" />
+			<input type="hidden" name="candidateId" value="${THE_CANDIDATE.id}" />
 			
 			<label class="label-form" style="margin-left:30px" >First Name</label>
-			<input type="text" name="name" value="${THE_STUDENT.name}" />
+			<input type="text" name="name" value="${THE_CANDIDATE.name}" />
 			
 			<label class="label-form" style="margin-left:30px">Last Name</label>
-			<input type="text" name="lastName" value="${THE_STUDENT.lastName}" />
+			<input type="text" name="lastName" value="${THE_CANDIDATE.lastName}" />
 			
 			<label class="label-form">E-mail</label>
-			<input type="text" name="email" value="${THE_STUDENT.email}" />
+			<input type="text" name="email" value="${THE_CANDIDATE.email}" />
 			
 			<label class="label-form">Phone</label>
-			<input type="text" name="phone" value="${THE_STUDENT.phone}" />
+			<input type="text" name="phone" value="${THE_CANDIDATE.phone}" />
 			
 			<label class="label-form" style="margin-left:35px">Department</label>
-			<input type="text" name="department" value="${THE_STUDENT.department}" />
+			<input type="text" name="department" value="${THE_CANDIDATE.department}" />
 			
 			<label class="label-form">Degree</label>
-			<input type="text" name="degree" value="${THE_STUDENT.degree}" />
+			<input type="text" name="degree" value="${THE_CANDIDATE.degree}" />
 			
 			<input type="submit" value="Save" class="save" />
 			<a class="btn-back" href="CandidateControllerServlet">Back to List</a>
